Smit

+++ server/data/basket.js
@@ -10,23 +10,48 @@ const {
   assertRequiredNumber,
 } = require('../utils/assertion');
 
-const { getUser } = require('./user');
+const usersData = require('./user');
+const { getGroup } = require('./group');
 
 const getByObjectId = async (objectId) => {
   const collection = await getBasketsCollection();
-  let cloth = await collection.findOne(idQuery(objectId));
-  return parseMongoData(cloth);
+  let basket = await collection.findOne(idQuery(objectId));
+  return parseMongoData(basket);
 };
 
 const addBasket = async (data) => {
   assertRequiredObject(data);
-  const { name, size, userId, groupId, users, clothes } = data;
+  const { name, size, userId, groupId, users, clothes, status, time } = data;
   const createdAt = new Date().getTime();
 
   assertObjectIdString(userId, 'Basket added by user ID');
   assertIsValuedString(name, 'Basket name');
   assertRequiredNumber(size, 'Basket size');
   assertIsValuedString(groupId, 'Group Id');
+  assertIsValuedString(status, 'Basket status');
+
+  if (
+    status != 'PENDING' &&
+    status != 'WASHING' &&
+    status != 'WASHING_DONE' &&
+    status != 'DRYING' &&
+    status != 'DRYING_DONE'
+  ) {
+    throw new QueryError(`Please provide proper status`);
+  }
+  assertRequiredNumber(time, 'Time');
+
+  const user = await usersData.getByObjectId(userId);
+  if (!user) {
+    throw new QueryError(`User not exist for user id (${userId})`);
+  }
+  const group = await getGroup(groupId);
+  if (!group) {
+    throw new QueryError(`Group not exist for group id (${groupId})`);
+  }
+  if (size < clothes.length) {
+    throw new QueryError(`No. of clothes must be less than size of basket.`);
+  }
 
   data.basketId = new ObjectId().toHexString();
 
@@ -34,11 +59,13 @@ const addBasket = async (data) => {
 
   const basketData = {
     _id: new ObjectId(data.basketId),
-    //groupId: new ObjectId(groupId),
+    groupId: new ObjectId(groupId),
     name,
     size,
     users,
     clothes,
+    status,
+    time,
     createdAt,
     updatedAt: createdAt,
     createdBy: new ObjectId(userId),
@@ -53,8 +80,13 @@ const addBasket = async (data) => {
   return basket;
 };
 
-const getBasket = async (id) => {
+const getBasket = async (userId, id) => {
   assertObjectIdString(id, 'Basket Id');
+  assertObjectIdString(userId, 'User Id');
+  const user = await usersData.getByObjectId(userId);
+  if (!user) {
+    throw new QueryError(`User not exist for user id (${userId})`);
+  }
   let basket = await getByObjectId(id);
   if (basket == null) {
     throw new QueryError(`Could not get basket for (${id})`);
@@ -62,48 +94,40 @@ const getBasket = async (id) => {
   return basket;
 };
 
-const getBasketByUserId = async (userId) => {
-  assertObjectIdString(userId, 'User Id');
-  // const user = await getUser(userId);
-  // console.log(user);
-  // if (!user) {
-  //   throw new QueryError(`Could not get user for (${userId})`);
-  // }
-
-  const collection = await getBasketsCollection();
-  let basket = [];
-  if (created) {
-    basket = await collection.find({ createdId: new ObjectId(userId) }).toArray();
-  } else {
-    basket = await collection.find({ 'baskets.users.id': new ObjectId(userId) });
-  }
-
-  if (basket == null) {
-    throw new QueryError(`Could not get basket for (${userId})`);
-  }
-  return basket;
-};
-
-const getBasketByGroupId = async (groupId) => {
+const getBasketByGroupId = async (userId, groupId) => {
   assertObjectIdString(groupId, 'Group Id');
-  // const user = await getUser(userId);
-  // console.log(user);
-  // if (!user) {
-  //   throw new QueryError(`Could not get user for (${userId})`);
-  // }
+  const user = await usersData.getByObjectId(userId);
+  if (!user) {
+    throw new QueryError(`User not exist for user id (${userId})`);
+  }
+
+  const group = await getGroup(groupId);
+  if (!group) {
+    throw new QueryError(`Group not exist for group id (${groupId})`);
+  }
 
   const collection = await getBasketsCollection();
-  let basket = [];
-  basket = await collection.find({ groupId: new ObjectId(groupId) }).toArray();
+
+  let basket = await collection.find({ groupId: new ObjectId(groupId) }).toArray();
 
   if (basket == null) {
     throw new QueryError(`Could not get basket for (${groupId})`);
   }
-  return basket;
+  let total = await collection.find({ groupId: new ObjectId(groupId) }).count();
+  return { basket, total };
 };
 
-const deleteBasket = async (id) => {
+const deleteBasket = async (userId, id) => {
   assertObjectIdString(id, 'Cloth id');
+  assertObjectIdString(userId, 'User Id');
+  const user = await usersData.getByObjectId(userId);
+  if (!user) {
+    throw new QueryError(`User not exist for user id (${userId})`);
+  }
+  let basket = await getByObjectId(id);
+  if (basket == null) {
+    throw new QueryError(`Could not get basket for (${id})`);
+  }
 
   let deleteBasket = await getBasket(id);
 
@@ -121,20 +145,18 @@ const deleteBasket = async (id) => {
   return deleteBasket;
 };
 
-const deleteBasketByUserId = async (userId) => {
-  const allBaskets = await getBasketByUserId(userId);
-
-  const collection = await getBasketsCollection();
-  const { deletedCount } = await collection.deleteMany({ createdBy: new ObjectId(userId) });
-  if (deletedCount === 0) {
-    throw new QueryError(`Could not delete all clothes for (${userId})`);
+const deleteBasketByGroupId = async (userId, groupId) => {
+  assertObjectIdString(id, 'Basket Id');
+  assertObjectIdString(userId, 'User Id');
+  const user = await usersData.getByObjectId(userId);
+  if (!user) {
+    throw new QueryError(`User not exist for user id (${userId})`);
   }
-  allBaskets.message = 'Successfully deleted all clothes.';
-  return parseMongoData(allBaskets);
-};
-
-const deleteBasketByGroupId = async (userId) => {
-  const allBaskets = await getBasketByGroupId(userId);
+  const group = await getGroup(groupId);
+  if (!group) {
+    throw new QueryError(`Group not exist for group id (${userId})`);
+  }
+  const allBaskets = await getBasketByGroupId(groupId);
 
   const collection = await getBasketsCollection();
   const { deletedCount } = await collection.deleteMany({ groupId: new ObjectId(groupId) });
@@ -145,12 +167,81 @@ const deleteBasketByGroupId = async (userId) => {
   return parseMongoData(allBaskets);
 };
 
+const updateBasket = async (id, data) => {
+  assertRequiredObject(data);
+  const { name, size, userId, groupId, users, clothes, status, time } = data;
+  const updatedAt = new Date().getTime();
+
+  assertObjectIdString(id, 'Basket ID');
+  assertObjectIdString(userId, 'Basket added by user ID');
+  assertIsValuedString(name, 'Basket name');
+  assertRequiredNumber(size, 'Basket size');
+  assertIsValuedString(groupId, 'Group Id');
+  assertIsValuedString(status, 'Basket status');
+
+  let previousBasket = await getByObjectId(id);
+
+  if (!previousBasket) throw new QueryError(`Could not found basket for (${id})`);
+
+  if (
+    status != 'PENDING' &&
+    status != 'WASHING' &&
+    status != 'WASHING_DONE' &&
+    status != 'DRYING' &&
+    status != 'DRYING_DONE'
+  ) {
+    throw new QueryError(`Please provide proper status`);
+  }
+  assertRequiredNumber(time, 'Time');
+
+  const user = await usersData.getByObjectId(userId);
+  if (!user) {
+    throw new QueryError(`User not exist for user id (${userId})`);
+  }
+  const group = await getGroup(groupId);
+  if (!group) {
+    throw new QueryError(`Group not exist for group id (${groupId})`);
+  }
+  if (size < clothes.length) {
+    throw new QueryError(`No. of clothes must be less than size of basket.`);
+  }
+
+  data.basketId = new ObjectId().toHexString();
+
+  const collection = await getBasketsCollection();
+
+  const basketData = {
+    groupId: new ObjectId(groupId),
+    name,
+    size,
+    users,
+    clothes,
+    status,
+    time,
+    createdAt: previousBasket.createdAt,
+    updatedAt,
+    createdBy: previousBasket.createdBy,
+    updatedBy: new ObjectId(userId),
+  };
+
+  const { modifiedCount, matchedCount } = await collection.updateOne(
+    { _id: new ObjectId(id), groupId: new ObjectId(groupId) },
+    { $set: basketData },
+  );
+
+  if (!modifiedCount && !matchedCount) {
+    throw new QueryError(`Could not update basket ID(${id})`);
+  }
+
+  const basket = await getByObjectId(id);
+  return basket;
+};
+
 module.exports = {
   addBasket,
   getBasket,
-  getBasketByUserId,
   getBasketByGroupId,
   deleteBasket,
-  deleteBasketByUserId,
   deleteBasketByGroupId,
+  updateBasket,
 };
